GCC compile fix.

+++ modules/Render/src/GPUProfileContext.c
@@ -101,7 +101,7 @@ static bool commandBufferValid(const dsCommandBuffer* commandBuffer)
 static uint32_t getHashTableSize(uint32_t querySize)
 {
 	const float loadFactor = 0.75f;
-	return (uint32_t)(querySize/loadFactor);
+	return (uint32_t)((float)querySize/loadFactor);
 }
 
 static uint32_t queryHash(const void* key)
